+++ karma-browsers.js
@@ -1,130 +1,109 @@
-module.exports = {
-    "IE 11.0 on Windows 10 r1024x768": {
-        "os": "Windows",
-        "os_version": "10",
-        "browser": "IE",
-        "browser_version": "11.0",
-        "resolution": "1024x768"
-    },
-    "Edge 14.0 on Windows 10 r1024x768": {
-        "os": "Windows",
-        "os_version": "10",
-        "browser": "Edge",
-        "browser_version": "14.0",
-        "resolution": "1024x768"
-    },
-    "Firefox 53.0 beta on Windows 10 r1024x768": {
-        "os": "Windows",
-        "os_version": "10",
-        "browser": "Firefox",
-        "browser_version": "53.0 beta",
-        "resolution": "1024x768"
-    },
-    "Chrome 56.0 on Windows 10 r1024x768": {
-        "os": "Windows",
-        "os_version": "10",
-        "browser": "Chrome",
-        "browser_version": "56.0",
-        "resolution": "1024x768"
-    },
-    "Safari 10.0 on OS X Sierra r1024x768": {
-        "os": "OS X",
-        "os_version": "Sierra",
-        "browser": "Safari",
-        "browser_version": "10.0",
-        "resolution": "1024x768"
-    },
-    "Firefox 53.0 beta on OS X Sierra r1024x768": {
-        "os": "OS X",
-        "os_version": "Sierra",
-        "browser": "Firefox",
-        "browser_version": "53.0 beta",
-        "resolution": "1024x768"
-    },
-    "Chrome 56.0 on OS X Sierra r1024x768": {
-        "os": "OS X",
-        "os_version": "Sierra",
-        "browser": "Chrome",
-        "browser_version": "56.0",
-        "resolution": "1024x768"
-    },
-    "Opera 12.15 on OS X Sierra r1024x768": {
-        "os": "OS X",
-        "os_version": "Sierra",
-        "browser": "Opera",
-        "browser_version": "12.15",
-        "resolution": "1024x768"
-    },
-    "iPhone 6S": {
-        "os": "ios",
-        "os_version": "9.1",
-        "browser": "iPhone",
-        "device": "iPhone 6S"
-    },
-    "iPhone 6S Plus": {
-        "os": "ios",
-        "os_version": "9.1",
-        "browser": "iPhone",
-        "device": "iPhone 6S Plus"
-    },
-    "iPad Air 2": {
-        "os": "ios",
-        "os_version": "9.1",
-        "browser": "iPad",
-        "device": "iPad Air 2"
-    },
-    "iPad Pro": {
-        "os": "ios",
-        "os_version": "9.1",
-        "browser": "iPad",
-        "device": "iPad Pro"
-    },
-    "iPad Mini 4": {
-        "os": "ios",
-        "os_version": "9.1",
-        "browser": "iPad",
-        "device": "iPad Mini 4"
-    },
-    "Samsung Galaxy S5 Mini": {
-        "os": "android",
-        "os_version": "4.4",
-        "browser": "Android Browser",
-        "device": "Samsung Galaxy S5 Mini"
-    },
-    "Samsung Galaxy S5": {
-        "os": "android",
-        "os_version": "4.4",
-        "browser": "Android Browser",
-        "device": "Samsung Galaxy S5"
-    },
-    "HTC One M8": {
-        "os": "android",
-        "os_version": "4.4",
-        "browser": "Android Browser",
-        "device": "HTC One M8"
-    },
-    "Motorola Razr Maxx HD": {
-        "os": "android",
-        "os_version": "4.0",
-        "browser": "Android Browser",
-        "device": "Motorola Razr"
-    },
-    "Amazon Kindle Fire HDX 7": {
-        "os": "android",
-        "os_version": "4.3",
-        "browser": "Android Browser",
-        "device": "Amazon Kindle Fire HDX 7"
-    },
-    "Samsung Galaxy Tab 4 10.1": {
-        "os": "android",
-        "os_version": "4.4",
-        "browser": "Android Browser",
-        "device": "Samsung Galaxy Tab 4 10.1"
-    }/*,
-    "Google Nexus 9": {
-        "os": "android",
-        "os_version": "5.0",
-        "browser": "Android Browser",
-        "device": "Google Nexus 9"
-    }*/
-};
+module.exports = [{
+    "os": "Windows",
+    "os_version": "10",
+    "browser": "IE",
+    "browser_version": "11.0",
+    "resolution": "1024x768"
+}, {
+    "os": "Windows",
+    "os_version": "10",
+    "browser": "Edge",
+    "browser_version": "14.0",
+    "resolution": "1024x768"
+}, {
+    "os": "Windows",
+    "os_version": "10",
+    "browser": "Firefox",
+    "browser_version": "53.0 beta",
+    "resolution": "1024x768"
+}, {
+    "os": "Windows",
+    "os_version": "10",
+    "browser": "Chrome",
+    "browser_version": "56.0",
+    "resolution": "1024x768"
+}, {
+    "os": "OS X",
+    "os_version": "Sierra",
+    "browser": "Safari",
+    "browser_version": "10.0",
+    "resolution": "1024x768"
+}, {
+    "os": "OS X",
+    "os_version": "Sierra",
+    "browser": "Firefox",
+    "browser_version": "53.0 beta",
+    "resolution": "1024x768"
+}, {
+    "os": "OS X",
+    "os_version": "Sierra",
+    "browser": "Chrome",
+    "browser_version": "56.0",
+    "resolution": "1024x768"
+}, {
+    "os": "OS X",
+    "os_version": "Sierra",
+    "browser": "Opera",
+    "browser_version": "12.15",
+    "resolution": "1024x768"
+}, {
+    "os": "ios",
+    "os_version": "9.1",
+    "browser": "iPhone",
+    "device": "iPhone 6S"
+}, {
+    "os": "ios",
+    "os_version": "9.1",
+    "browser": "iPhone",
+    "device": "iPhone 6S Plus"
+}, {
+    "os": "ios",
+    "os_version": "9.1",
+    "browser": "iPad",
+    "device": "iPad Air 2"
+}, {
+    "os": "ios",
+    "os_version": "9.1",
+    "browser": "iPad",
+    "device": "iPad Pro"
+}, {
+    "os": "ios",
+    "os_version": "9.1",
+    "browser": "iPad",
+    "device": "iPad Mini 4"
+}, {
+    "os": "android",
+    "os_version": "4.4",
+    "browser": "Android Browser",
+    "device": "Samsung Galaxy S5 Mini"
+}, {
+    "os": "android",
+    "os_version": "4.4",
+    "browser": "Android Browser",
+    "device": "Samsung Galaxy S5"
+}, {
+    "os": "android",
+    "os_version": "4.4",
+    "browser": "Android Browser",
+    "device": "HTC One M8"
+}, {
+    "os": "android",
+    "os_version": "4.0",
+    "browser": "Android Browser",
+    "device": "Motorola Razr"
+}, {
+    "os": "android",
+    "os_version": "4.3",
+    "browser": "Android Browser",
+    "device": "Amazon Kindle Fire HDX 7"
+}, {
+    "os": "android",
+    "os_version": "4.4",
+    "browser": "Android Browser",
+    "device": "Samsung Galaxy Tab 4 10.1"
+}/*, {
+    "os": "android",
+    "os_version": "5.0",
+    "browser": "Android Browser",
+    "device": "Google Nexus 9"
+}*/];
