+++ package.json
@@ -2,30 +2,37 @@
   "name": "frame",
   "version": "0.0.0",
   "scripts": {
-    "dev": "webpack-dev-server --open --hot --host 0.0.0.0",
+    "start": "webpack-dev-server --open --hot --host 0.0.0.0",
     "build": "cross-env NODE_ENV=production webpack",
-    "devtest": "karma start",
-    "test": "karma start --no-auto-watch --single-run"
+    "test": "karma start"
   },
   "devDependencies": {
     "babel-core": "^6.0.0",
     "babel-loader": "^7.0.0",
     "babel-preset-env": "^1.4.0",
+    "copy-webpack-plugin": "^4.0.1",
     "cross-env": "^4.0.0",
     "css-loader": "^0.28.1",
+    "favicons-webpack-plugin": "0.0.7",
     "file-loader": "^0.11.0",
     "firebase": "^3.7.8",
+    "html-webpack-include-assets-plugin": "0.0.5",
+    "html-webpack-plugin": "^2.28.0",
     "karma": "^1.6.0",
     "karma-browserstack-launcher": "^1.2.0",
+    "karma-chrome-launcher": "^2.1.1",
     "karma-firebase-server": "^0.1.1",
     "karma-mocha": "^1.3.0",
-    "karma-phantomjs-launcher": "^1.0.4",
+    "karma-mocha-reporter": "^2.2.3",
+    "karma-sourcemap-loader": "^0.3.7",
     "karma-webpack": "^2.0.3",
+    "loader-utils": "^1.1.0",
     "material-components-web": "^0.10.0",
     "mocha": "^3.3.0",
     "node-sass": "^4.5.0",
     "sass-loader": "^6.0.0",
     "three": "^0.85.2",
+    "url-loader": "^0.5.8",
     "uuid": "^3.0.1",
     "vue": "^2.3.2",
     "vue-loader": "^12.0.3",
