--- src/index.html
@@ -1,7 +1,0 @@
-<!DOCTYPE html>
-<meta charset="utf-8">
-<link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
-<link href="https://fonts.googleapis.com/css?family=Roboto:300,400,500" rel="stylesheet">
-<title>Frame | 1stop-st.org</title>
-<div></div>
-<script src="build.js"></script>
