+++ src/vue-component/canvas/class-lines.js
