+++ src/vue-component/canvas/class-nodes.js
