+++ src/vue-component/canvas/index.vue
