+++ src/vue-component/fab/index.vue
