+++ src/vue-component/splitter/index.vue
