+++ src/vuex-store/modules/database-connection/getters.js
