+++ test/test_src/test_vuex-store/test_modules/test_database-connection/test_model-sample.js
@@ -1,0 +1,589 @@
+export let data = {
+    "initdata": {
+        "input": {
+            "boundaries": {
+                "a": {
+                    "node": 0,
+                    "rx": true,
+                    "ry": true,
+                    "rz": true,
+                    "x": true,
+                    "y": true,
+                    "z": true
+                },
+                "b": {
+                    "node": 1,
+                    "rx": true,
+                    "ry": true,
+                    "rz": true,
+                    "x": true,
+                    "y": true,
+                    "z": true
+                },
+                "c": {
+                    "node": 2,
+                    "rx": true,
+                    "ry": true,
+                    "rz": true,
+                    "x": true,
+                    "y": true,
+                    "z": true
+                },
+                "d": {
+                    "node": 4,
+                    "rx": true,
+                    "ry": true,
+                    "rz": true,
+                    "x": true,
+                    "y": true,
+                    "z": true
+                },
+                "e": {
+                    "node": 3,
+                    "rx": true,
+                    "ry": true,
+                    "rz": true,
+                    "x": 0,
+                    "y": 0,
+                    "z": 0
+                },
+                "f": {
+                    "node": 5,
+                    "rx": true,
+                    "ry": true,
+                    "rz": true,
+                    "x": 0,
+                    "y": 0,
+                    "z": 0
+                },
+                "g": {
+                    "node": 6,
+                    "rx": true,
+                    "ry": true,
+                    "rz": true,
+                    "x": 0,
+                    "y": 0,
+                    "z": 0
+                },
+                "h": {
+                    "node": 7,
+                    "rx": true,
+                    "ry": true,
+                    "rz": true,
+                    "x": 0,
+                    "y": 0,
+                    "z": 0
+                }
+            },
+            "lines": {
+                "a": {
+                    "EA": 1.0,
+                    "material": 0,
+                    "n1": 0,
+                    "n2": 3,
+                    "section": 0
+                },
+                "b": {
+                    "EA": 1.0,
+                    "material": 0,
+                    "n1": 1,
+                    "n2": 5,
+                    "section": 0
+                },
+                "c": {
+                    "EA": 1.0,
+                    "material": 0,
+                    "n1": 2,
+                    "n2": 6,
+                    "section": 0
+                },
+                "d": {
+                    "EA": 1.0,
+                    "material": 0,
+                    "n1": 3,
+                    "n2": 5,
+                    "section": 0
+                },
+                "e": {
+                    "EA": 1.0,
+                    "material": 0,
+                    "n1": 3,
+                    "n2": 6,
+                    "section": 0
+                },
+                "f": {
+                    "EA": 1.0,
+                    "material": 0,
+                    "n1": 4,
+                    "n2": 7,
+                    "section": 0
+                },
+                "g": {
+                    "EA": 1.0,
+                    "material": 0,
+                    "n1": 5,
+                    "n2": 7,
+                    "section": 0
+                },
+                "h": {
+                    "EA": 1.0,
+                    "material": 0,
+                    "n1": 6,
+                    "n2": 7,
+                    "section": 0
+                },
+                "i": {
+                    "EA": 1.0,
+                    "material": 0,
+                    "n1": 0,
+                    "n2": 5,
+                    "section": 0
+                },
+                "j": {
+                    "EA": 1.0,
+                    "material": 0,
+                    "n1": 4,
+                    "n2": 5,
+                    "section": 0
+                },
+                "k": {
+                    "EA": 1.0,
+                    "material": 0,
+                    "n1": 0,
+                    "n2": 6,
+                    "section": 0
+                },
+                "l": {
+                    "EA": 1.0,
+                    "material": 0,
+                    "n1": 4,
+                    "n2": 6,
+                    "section": 0
+                },
+                "m": {
+                    "EA": 1.0,
+                    "material": 0,
+                    "n1": 5,
+                    "n2": 6,
+                    "section": 0
+                },
+                "n": {
+                    "EA": 1.0,
+                    "material": 0,
+                    "n1": 3,
+                    "n2": 1,
+                    "section": 0
+                },
+                "o": {
+                    "EA": 1.0,
+                    "material": 0,
+                    "n1": 2,
+                    "n2": 7,
+                    "section": 0
+                },
+                "p": {
+                    "EA": 1.0,
+                    "material": 0,
+                    "n1": 2,
+                    "n2": 3,
+                    "section": 0
+                },
+                "q": {
+                    "EA": 1.0,
+                    "material": 0,
+                    "n1": 1,
+                    "n2": 7,
+                    "section": 0
+                },
+                "r": {
+                    "EA": 1.0,
+                    "material": 0,
+                    "n1": 3,
+                    "n2": 7,
+                    "section": 0
+                }
+            },
+            "materials": {
+                "a": {
+                    "E": 1,
+                    "G": 1
+                }
+            },
+            "nodeLoads": {
+                "a": {
+                    "node": 3,
+                    "rx": 0.0,
+                    "ry": 0.0,
+                    "rz": 0.0,
+                    "x": 1.0,
+                    "y": 0.0,
+                    "z": 0.0
+                },
+                "b": {
+                    "node": 6,
+                    "rx": 0.0,
+                    "ry": 0.0,
+                    "rz": 0.0,
+                    "x": 1.0,
+                    "y": 0.0,
+                    "z": 0.0
+                }
+            },
+            "nodes": {
+                "a": {
+                    "x": 0,
+                    "y": 0.0,
+                    "z": 0.0
+                },
+                "b": {
+                    "x": 3.0,
+                    "y": 0.0,
+                    "z": 0.0
+                },
+                "c": {
+                    "x": 0.0,
+                    "y": 3.0,
+                    "z": 0.0
+                },
+                "d": {
+                    "x": 0.0,
+                    "y": 0.0,
+                    "z": 3.0
+                },
+                "e": {
+                    "x": 3.0,
+                    "y": 3.0,
+                    "z": 0.0
+                },
+                "f": {
+                    "x": 3.0,
+                    "y": 0.0,
+                    "z": 3.0
+                },
+                "g": {
+                    "x": 0.0,
+                    "y": 3.0,
+                    "z": 3.0
+                },
+                "h": {
+                    "x": 3.0,
+                    "y": 3.0,
+                    "z": 3.0
+                }
+            },
+            "sections": {
+                "a": {
+                    "B": 3.125,
+                    "H": 3.2,
+                    "shape": "H",
+                    "tf": 0.1,
+                    "tw": 0.125
+                }
+            }
+        }
+    },
+    "userdata": {
+        "08RoeVSUR3OLSeHp1qWUQIlAiTz1": {
+            "filenames": {
+                "abedkljr": "Untitled1"
+            },
+            "input": {
+                "abedkljr": {
+                    "boundaries": {
+                        "a": {
+                            "node": 0,
+                            "rx": true,
+                            "ry": true,
+                            "rz": true,
+                            "x": true,
+                            "y": true,
+                            "z": true
+                        },
+                        "b": {
+                            "node": 1,
+                            "rx": true,
+                            "ry": true,
+                            "rz": true,
+                            "x": true,
+                            "y": true,
+                            "z": true
+                        },
+                        "c": {
+                            "node": 2,
+                            "rx": true,
+                            "ry": true,
+                            "rz": true,
+                            "x": true,
+                            "y": true,
+                            "z": true
+                        },
+                        "d": {
+                            "node": 4,
+                            "rx": true,
+                            "ry": true,
+                            "rz": true,
+                            "x": true,
+                            "y": true,
+                            "z": true
+                        },
+                        "e": {
+                            "node": 3,
+                            "rx": true,
+                            "ry": true,
+                            "rz": true,
+                            "x": 0,
+                            "y": 0,
+                            "z": 0
+                        },
+                        "f": {
+                            "node": 5,
+                            "rx": true,
+                            "ry": true,
+                            "rz": true,
+                            "x": 0,
+                            "y": 0,
+                            "z": 0
+                        },
+                        "g": {
+                            "node": 6,
+                            "rx": true,
+                            "ry": true,
+                            "rz": true,
+                            "x": 0,
+                            "y": 0,
+                            "z": 0
+                        },
+                        "h": {
+                            "node": 7,
+                            "rx": true,
+                            "ry": true,
+                            "rz": true,
+                            "x": 0,
+                            "y": 0,
+                            "z": 0
+                        }
+                    },
+                    "lines": {
+                        "a": {
+                            "EA": 1.0,
+                            "material": 0,
+                            "n1": 0,
+                            "n2": 3,
+                            "section": 0
+                        },
+                        "b": {
+                            "EA": 1.0,
+                            "material": 0,
+                            "n1": 1,
+                            "n2": 5,
+                            "section": 0
+                        },
+                        "c": {
+                            "EA": 1.0,
+                            "material": 0,
+                            "n1": 2,
+                            "n2": 6,
+                            "section": 0
+                        },
+                        "d": {
+                            "EA": 1.0,
+                            "material": 0,
+                            "n1": 3,
+                            "n2": 5,
+                            "section": 0
+                        },
+                        "e": {
+                            "EA": 1.0,
+                            "material": 0,
+                            "n1": 3,
+                            "n2": 6,
+                            "section": 0
+                        },
+                        "f": {
+                            "EA": 1.0,
+                            "material": 0,
+                            "n1": 4,
+                            "n2": 7,
+                            "section": 0
+                        },
+                        "g": {
+                            "EA": 1.0,
+                            "material": 0,
+                            "n1": 5,
+                            "n2": 7,
+                            "section": 0
+                        },
+                        "h": {
+                            "EA": 1.0,
+                            "material": 0,
+                            "n1": 6,
+                            "n2": 7,
+                            "section": 0
+                        },
+                        "i": {
+                            "EA": 1.0,
+                            "material": 0,
+                            "n1": 0,
+                            "n2": 5,
+                            "section": 0
+                        },
+                        "j": {
+                            "EA": 1.0,
+                            "material": 0,
+                            "n1": 4,
+                            "n2": 5,
+                            "section": 0
+                        },
+                        "k": {
+                            "EA": 1.0,
+                            "material": 0,
+                            "n1": 0,
+                            "n2": 6,
+                            "section": 0
+                        },
+                        "l": {
+                            "EA": 1.0,
+                            "material": 0,
+                            "n1": 4,
+                            "n2": 6,
+                            "section": 0
+                        },
+                        "m": {
+                            "EA": 1.0,
+                            "material": 0,
+                            "n1": 5,
+                            "n2": 6,
+                            "section": 0
+                        },
+                        "n": {
+                            "EA": 1.0,
+                            "material": 0,
+                            "n1": 3,
+                            "n2": 1,
+                            "section": 0
+                        },
+                        "o": {
+                            "EA": 1.0,
+                            "material": 0,
+                            "n1": 2,
+                            "n2": 7,
+                            "section": 0
+                        },
+                        "p": {
+                            "EA": 1.0,
+                            "material": 0,
+                            "n1": 2,
+                            "n2": 3,
+                            "section": 0
+                        },
+                        "q": {
+                            "EA": 1.0,
+                            "material": 0,
+                            "n1": 1,
+                            "n2": 7,
+                            "section": 0
+                        },
+                        "r": {
+                            "EA": 1.0,
+                            "material": 0,
+                            "n1": 3,
+                            "n2": 7,
+                            "section": 0
+                        }
+                    },
+                    "materials": {
+                        "a": {
+                            "E": 1,
+                            "G": 1
+                        }
+                    },
+                    "nodeLoads": {
+                        "a": {
+                            "node": 3,
+                            "rx": 0.0,
+                            "ry": 0.0,
+                            "rz": 0.0,
+                            "x": 1.0,
+                            "y": 0.0,
+                            "z": 0.0
+                        },
+                        "b": {
+                            "node": 6,
+                            "rx": 0.0,
+                            "ry": 0.0,
+                            "rz": 0.0,
+                            "x": 1.0,
+                            "y": 0.0,
+                            "z": 0.0
+                        }
+                    },
+                    "nodes": {
+                        "a": {
+                            "x": 0,
+                            "y": 0.0,
+                            "z": 0.0
+                        },
+                        "b": {
+                            "x": 3.0,
+                            "y": 0.0,
+                            "z": 0.0
+                        },
+                        "c": {
+                            "x": 0.0,
+                            "y": 3.0,
+                            "z": 0.0
+                        },
+                        "d": {
+                            "x": 0.0,
+                            "y": 0.0,
+                            "z": 3.0
+                        },
+                        "e": {
+                            "x": 3.0,
+                            "y": 3.0,
+                            "z": 0.0
+                        },
+                        "f": {
+                            "x": 3.0,
+                            "y": 0.0,
+                            "z": 3.0
+                        },
+                        "g": {
+                            "x": 0.0,
+                            "y": 3.0,
+                            "z": 3.0
+                        },
+                        "h": {
+                            "x": 3.0,
+                            "y": 3.0,
+                            "z": 3.0
+                        }
+                    },
+                    "sections": {
+                        "a": {
+                            "B": 3.125,
+                            "H": 3.2,
+                            "shape": "H",
+                            "tf": 0.1,
+                            "tw": 0.125
+                        }
+                    }
+                }
+            }
+        }
+    }
+};
+
+export let rule = {
+    "rules": {
+        "userdata": {
+            "$uid": {
+                ".read": "auth.uid === $uid",
+                ".write": "auth.uid === $uid"
+            }
+        },
+        "initdata": {
+            ".read": true
+        }
+    }
+};
